requirements.txt
dnspython==2.2.1
mysql==0.0.3
mysqlclient==2.1.1
pymongo==4.3.2
python-dotenv==0.21.0
